Another attempt at error handling
Scattered comments online suggest that req and res need error handling,
so I'll try adding error handling there.

Use console.log to show a message, then console.log to show the error
object (so that it can inspect it).  This should be better than
console.error (because web errors may be expected) and better than
showing the error in a template string (which will just print `[object
Object]` or similar) and better than showing the error without context
or explanation.