Properly implement iOS (HTTPS) support
FFRK was apparently aborting the connection whenever it connected.
Comparison with Fiddler (which did work) showed some differences in
the encryption algorithm (SHA-256, instead of SHA-1 in the Forge example
code) and the handling of subjectAltName, so I made those changes.

Update proxy logic to properly and fully intercept and proxy FFRK
requests.  This included tweaking proxy logic to handle HTTPS versus
HTTP and updating the startup detection to handle both HTTPS and HTTP.