Properly functioning tooltips
Rendering the tooltip component alongside the cell didn't work, because
ag-grid + React wouldn't necessarily mount the components in an order
such that the tooltip could see its corresponding cell.

And, besides, that's probably needlessly slow.

Instead, render one grid-level tooltip, and use the data-tip attribute
to pass in the ID of the cell to render.

Upgrade react-tooltip so that I can pass parameters to getContent.

Set PrizeList's bottom margin to 0.  In practice, it was 0 where it was
previously used anyway (due to Bootstrap's or reboot's styling), and we
want it to be 0 for the Torment grid tooltips, so...