Rework stat mod handling
Change how status effects are broken up to try to better handle
ambiguities such as whether durations and "who"s apply to the
immediately preceding terms or all terms.

Let all "grants" and "causes" stat mods follow the status effect code
path, instead of trying to process them as stat mods.  Prior to doing
this, it was too hard to correctly process grants-and-causes statuses
versus grants-and-causes stat mods versus non-grants-and-causes stat
mods.  (Auron's Law Breaker was a particular challenge: "causes DEF, RES
and MND -70% for 8 seconds" resulted in "DEF" getting processed as a
separate status, because "RES and MND" looked like a stat mod and thus
the end of the status effect processing.)