Fix configureStore for browser; fix styles; verbose output
Add an IS_ELECTRON environment variable and use that to configure the
store so that it works in a web browser.

Turn off noInfo for webpack.config.renderer.dev.js to have more verbose
output.

Add Bootstrap stylesheet for public index.html.  Delete unnecessary
comments.

Fix a styling glitch - the side nav took up the page height in Chrome
but not in Electron.  I do not know why.