Fix webpack builds
node_modules weren't being included, so both the renderer and the main
process weren't working.

After my changes, node modules are getting directly bundled with both
the renderer and the main process.  This is probably somewhat
inefficient, because it probably means that some of the same code is
included with both.  However, it's still a lot better than bundling all
of node_modules (including all of Font Awesome, etc.), and the bundled
modules seem to compress well.