Experimental fix for Lunarian Might and similar mixed neg+pos statuses
The previous code handled this with an awkward negative lookahead clause
within its lookaheadWho logic to try to exclude who clauses that were
followed by durations.  However, I believe that was simply a hack
instead of something that necessarily universally worked.

The new code instead tries to assign the target as the who of the first
status.  This may not work, either, depending on how many statuses are
present, but it's worth a try.