Allow arbitrary standalone attack extras
This lets us handle Waiting for Friends as written:

> Ten single attacks (0.49 each), grants Soul Break Gauge +250, 100%
> additional critical chance, 100% hit rate, removes Waiting for Friends
> Mode

instead of having to edit it so that "100% additional critical chance"
goes with the attack - that may technically make more sense, but being
more tolerant of (reasonable) forms simplifies maintenance.

Show 0 multiplier attacks as "0.0" instead of "0.0/0"; this is to
accommodate Exdeath's SASB finisher.