Don't log timestamps for enlir*-to-mrp scripts
This makes it easier to diff and inspect their contents.

Delete the `scripts` symlinks to logging files.  They were added to make it easier to share code with Luca, but Luca is abandoned, and RK-Squared is open source, and they cause problems such as confusing WebStorm and causing the modules to be loaded twice.