Code cleanup
Delete expandStatusSlashOptions.  That was used to accommodate ambiguities like "Do Something/Poison", but it's not needed now that status names are clearly delineated.

Delete checkForAndStatuses (the "Beast and Father" handler).  It's also unneeded now that status names are clearly delineated.

Delete current logic for handling hybrid ("or") statuses and stat buffs. We need to do something for these, but what we currently have is broken.

Delete unused parsePercentageCounts and numberWithCommas.